Move abstract input parser class to submodule level.
And document.
